+++ changelog.md
@@ -7,6 +7,8 @@
     steemvoter to vote (Default: `90%`). Current voting power is updated via RPC every 20 seconds.
 * A new config key, `"backup_author_names"`, is used to specify authors that
     should be voted for when steemvoter is not using enough voting power.
+* A new config key, `"max_voting_power"`, is used to specify the remaining voting power
+    above which steemvoter will vote for backup authors.
 * A new config key, `"vote_interval"`, is used to specify the timespan that
     steemvoter will wait between collecting and voting on eligible comments (Default: `10 seconds`).
 * The config key `"vote_delay"` is now `"min_post_age"`. The effect is the same.